etl_pipeline/requirements.txt
boto3
pandas
mysql-connector-python